Added scaling feature to viewer.
git-svn-id: http://kos.informatik.uni-osnabrueck.de/software/meshing@2369 bea721bf-f9f1-0310-a2d5-e0bfdc8dd81e

+++ src/viewer/TouchPad.cpp
@@ -43,8 +43,8 @@ bool TouchPad::checkInput(int d){
 
 void TouchPad::mouseMoveEvent(QMouseEvent* e){
 
-	int dx = old_x - e->x();
-	int dy = old_y - e->y();
+	double dx = (double)old_x - e->x();
+	double dy = (double)old_y - e->y();
 
 	switch(transformation_mode){
 	case YAW_ROLL:
